Harry Potter - Reading Mech
Only able to read & sort. It currently reads the text documents, it generates 80% training data and 20% testing and stores testing for later. 
Then it does a word count for each book, the word and count are in order with regards to the indices.
It also has one large book document containing all words from the training data and the word count.